Proof of concept that you can indeed sweep the tos bits
I twiddled with this back in january, trying to understand the
relationship between the iov, the cmsg, and the msg header structures
and I finally grok it. The internet was of no help. The kernel
support was obviously there, and also buried deep in the glibc
send and sendto calls....

However the answer is too long to fit into this commit and there
are still issues to sort out with ipv4 vs ipv6 vs ipv6mapped sockets.

But with this, we can, indeed, sweep the values of the dscp field on
a per packet basis.

d@puck:~/git/isochronous$ ./udpburst -f localhost -C
IPV6_RECVTCLASS: Protocol not available
controllen: 20
64 bytes -- received 32 of 32 -- 32 consecutive 0 ooo 0 dups 8 ect

 1 2 3 4 5 6 7 8 9 a b c d e f101112131415161718191a1b1c1d1e1f